refactor(PaymentConfirmRequest): 금액 type 수정

+++ src/main/java/roomescape/payment/dto/PaymentConfirmRequest.java
@@ -1,11 +1,12 @@
 package roomescape.payment.dto;
 
+import java.math.BigDecimal;
 import roomescape.reservation.dto.UserReservationCreateRequest;
 
 public record PaymentConfirmRequest(
         String paymentKey,
         String orderId,
-        long amount
+        BigDecimal amount
 ) {
     public static PaymentConfirmRequest from(UserReservationCreateRequest request) {
         return new PaymentConfirmRequest(request.paymentKey(), request.orderId(), request.amount());
